backend/python_model/requirements.txt
pandas
python-pptx
PyMuPDF
spacy
opencv-python
pytesseract
Pillow
openpyxl